requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.14.1
